Add json files data to the project
