+++ book.rb
@@ -1,9 +1,14 @@
 class Book
+  attr_accessor :title, :author, :rentals
+
   def initialize(title, author)
     @title = title,
-             @author = author
+    @author = author
     @rentals = []
   end
 
-  attr_accessor :title, :author, :rentals
+  def add_rental(person, date)
+    Rental.new(date, self, person)
+  end
+
 end
